Move providers to component.spi packages
