+++ api/src/test/java/io/opentelemetry/OpenTelemetryTest.java
@@ -30,19 +30,19 @@ import io.opentelemetry.metrics.Measure;
 import io.opentelemetry.metrics.Measurement;
 import io.opentelemetry.metrics.Meter;
 import io.opentelemetry.metrics.NoopMetrics;
+import io.opentelemetry.metrics.spi.MeterProvider;
 import io.opentelemetry.resource.Resource;
-import io.opentelemetry.spi.MeterProvider;
-import io.opentelemetry.spi.TaggerProvider;
-import io.opentelemetry.spi.TracerProvider;
 import io.opentelemetry.tags.NoopTags;
 import io.opentelemetry.tags.TagMap;
 import io.opentelemetry.tags.Tagger;
+import io.opentelemetry.tags.spi.TaggerProvider;
 import io.opentelemetry.trace.NoopTrace;
 import io.opentelemetry.trace.Span;
 import io.opentelemetry.trace.Span.Builder;
 import io.opentelemetry.trace.SpanContext;
 import io.opentelemetry.trace.SpanData;
 import io.opentelemetry.trace.Tracer;
+import io.opentelemetry.trace.spi.TracerProvider;
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
